+++ NEWS.md
@@ -1,6 +1,6 @@
 # Ternary v1.1.4.9002 (development)
 
-- Option `axis.rotate` to allow axis labels to be plotted 'unrotated' 
+- Optional parameter `axis.rotate` to allow axis labels to be plotted 'unrotated' 
    (https://github.com/ms609/Ternary/issues/38).
 
 # Ternary v1.1.4.9001 (development)